+++ app/src/main/java/com/zero/neon/game/enemy/ship/model/Boss.kt
@@ -81,7 +81,7 @@ data class Boss(
         )
     }
 
-    override fun onObjectImpact(impactPower: Int) {
+    override fun onObjectImpact(impactPower: Float) {
         hp -= impactPower
     }
